# Enable json v2 support for 59% faster sourcemap parsing